Small refactor and moved proxy tests to new test-file.
