+++ examples/WireMock.Net.Console.Record.NETCoreApp/WireMock.Net.Console.Record.NETCoreApp.csproj
@@ -15,4 +15,8 @@
     <PackageReference Include="Newtonsoft.Json" Version="10.0.1" />
   </ItemGroup>
 
+  <ItemGroup>
+    <Folder Include="__admin\mappings\" />
+  </ItemGroup>
+
 </Project>
